modified

+++ TRI.py
@@ -1,6 +1,6 @@
 import math
-a = int(input("Enter 1st no. :"))
-b = int(input("Enter 2st no. :"))
+a = float(input("Enter 1st no. :"))
+b = float(input("Enter 2st no. :"))
 s1 = math.sqrt(a)
 s2 = math.sqrt(b)
 print("square root of a=", s1)
